% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_alluvium.R
\name{alluvium_UI}
\alias{alluvium_UI}
\alias{alluvium_server}
\title{A module that produces a faceted line chart.}
\usage{
alluvium_UI(id)

alluvium_server(id, ds)
}
\arguments{
\item{ds}{A data.framish dataset. **Reactive**.}
}
\value{
Nothing
}
\description{
This module presents a faceted line chart of the variables present in the dataset.
}
\section{Functions}{
\itemize{
\item \code{alluvium_server}: Line Plot Server function
}}

