% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_bar.R
\name{bar_UI}
\alias{bar_UI}
\alias{bar_server}
\title{A module that produces a faceted line chart.}
\usage{
bar_UI(id)

bar_server(id, ds)
}
\arguments{
\item{ds}{A data.framish dataset. **Reactive**.}
}
\value{
Nothing
}
\description{
This module presents a faceted line chart of the variables present in the dataset.
}
\section{Functions}{
\itemize{
\item \code{bar_server}: Line Plot Server function
}}

