% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_flowbar.R
\name{flowbar_UI}
\alias{flowbar_UI}
\alias{flowbar_server}
\title{A module that produces a faceted line chart.}
\usage{
flowbar_UI(id)

flowbar_server(id, ds)
}
\arguments{
\item{ds}{A data.framish dataset. **Reactive**.}
}
\value{
Nothing
}
\description{
This module presents a faceted line chart of the variables present in the dataset.
}
\section{Functions}{
\itemize{
\item \code{flowbar_server}: Line Plot Server function
}}

