% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_line.R
\name{line_UI}
\alias{line_UI}
\alias{line_server}
\title{A module that produces a faceted line chart.}
\usage{
line_UI(id)

line_server(id, ds)
}
\arguments{
\item{ds}{A data.framish dataset. **Reactive**.}
}
\value{
Nothing
}
\description{
This module presents a faceted line chart of the variables present in the dataset.
}
\section{Functions}{
\itemize{
\item \code{line_server}: Line Plot Server function
}}

